requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf>=4.25.0,<7.0.0
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.23.5
